+++ .github/workflows/ci.yml
@@ -16,7 +16,7 @@ jobs:
 
     strategy:
       matrix:
-        go: [ '1.19.x' ]
+        go: [ '1.19.x', '1.20.x' ]
 
     steps:
       - name: Set up Go
