+++ go.mod
@@ -2,6 +2,9 @@ module github.com/essentialkaos/mdtoc
 
 go 1.17
 
-require github.com/essentialkaos/ek/v12 v12.60.0
+require (
+	github.com/essentialkaos/depsy v1.0.0
+	github.com/essentialkaos/ek/v12 v12.60.0
+)
 
 require golang.org/x/sys v0.5.0 // indirect
